+++ qgreenland/__init__.py
@@ -1,1 +1,1 @@
-__version__ = 'v2.0.0beta1'
+__version__ = 'v2.0.0beta2'